+++ lib/blocs/recipes/bloc.dart
@@ -2,7 +2,7 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:recipes/database.dart';
 import 'package:recipes/blocs/recipes/events.dart';
 import 'package:recipes/blocs/recipes/state.dart';
-import 'package:recipes/recipe.dart';
+import 'package:recipes/models/recipe.dart';
 
 class RecipesBloc extends Bloc<RecipesEvent, RecipesState> {
   final DatabaseClient databaseClient;
