+++ lib/blocs/recipes/events.dart
@@ -1,5 +1,5 @@
 import 'package:equatable/equatable.dart';
-import 'package:recipes/recipe.dart';
+import 'package:recipes/models/recipe.dart';
 
 abstract class RecipesEvent extends Equatable {
   const RecipesEvent();
