+++ lib/screens/edit_meal_plan.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:get_it/get_it.dart';
 import 'package:recipes/database.dart';
-import 'package:recipes/meal_plan.dart';
-import 'package:recipes/recipe.dart';
+import 'package:recipes/models/meal_plan.dart';
+import 'package:recipes/models/recipe.dart';
 import 'package:recipes/screens/meal_plan.dart';
 
 class EditMealPlan extends StatelessWidget {
