+++ lib/screens/groceries.dart
@@ -5,7 +5,7 @@ import 'package:recipes/blocs/groceries/bloc.dart';
 import 'package:recipes/blocs/groceries/events.dart';
 import 'package:recipes/blocs/groceries/state.dart';
 import 'package:recipes/database.dart';
-import 'package:recipes/grocery.dart';
+import 'package:recipes/models/grocery.dart';
 import 'package:recipes/helpers/ingredient_formatters.dart';
 
 class GroceriesList extends StatelessWidget {
